updates to the code after testing by Lola & Sara

+++ BuildME/energy.py
@@ -10,17 +10,18 @@ import platform
 from BuildME import settings
 
 
-def perform_energy_calculation(out_dir, ep_dir, epw_path):
+def perform_energy_calculation(out_dir, ep_dir, epw_path, keep_all):
     """
     Copies the required EnergyPlus files and initiates the energy demand simulation
     :param out_dir: output folder directory
     :param ep_dir: EnergyPlus directory
     :param epw_path: path to the EPW file with weather data
+    :param keep_all: boolean indicating whether to keep all simulation files (incl. the .eso file)
     """
-    print("Perform energy simulation...")
     copy_files(out_dir, ep_dir, epw_path)
     run_energyplus_single(out_dir)
-    delete_ep_files(out_dir)
+    if not keep_all:
+        delete_ep_files(out_dir)
     return
 
 
@@ -30,13 +31,15 @@ def perform_energy_calculation_mp(args):
     :param args: arguments (out_dir - output folder directory,
                             ep_dir - EnergyPlus directory,
                             epw_path - path to the EPW file with weather data,
+                            keep_all - boolean indicating whether to keep all simulation files (incl. the .eso file)
                             q - multiprocessing Queue object,
                             no - iteration number)
     """
-    out_dir, ep_dir, epw_path, q, no = args
+    out_dir, ep_dir, epw_path, keep_all, q, no = args
     copy_files(out_dir, ep_dir, epw_path)
     run_energyplus_single(out_dir)
-    delete_ep_files(out_dir)
+    if not keep_all:
+        delete_ep_files(out_dir)
     q.put(no)
     return
 
@@ -168,7 +171,7 @@ def run_energyplus_single(out_dir, verbose=True):
         run_idf = 'in.idf'
 
     with open("log_energyplus.txt", 'w+') as log_file:
-        cmd = out_dir + '/energyplus -r %s' % run_idf
+        cmd = f'"{os.path.join(out_dir, "energyplus")}" -r {run_idf}'
         log_file.write("%s\n\n" % cmd)
         log_file.flush()
         subprocess.call(cmd, shell=True, stdout=log_file, stderr=log_file)
